Integrate wind_offshore in eGon100RE
